+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+What's changed since pre-release v0.1.0-B2105002:
+
+- New features:
+  - Automatically export repository data when running in Github Actions. [#80](https://github.com/microsoft/PSRule.Rules.GitHub/issues/80)
+- Engineering:
+  - Bump PSRule dependency to v1.7.0. [#79](https://github.com/microsoft/PSRule.Rules.GitHub/issues/79)
+
 ## v0.1.0-B2105002 (pre-release)
 
 What's changed since pre-release v0.1.0-B2104014: